a7: refatorando a estrutura do projeto
exercicios resolvidos
