+++ fast_zero/app.py
@@ -1,23 +1,13 @@
-from http import HTTPStatus
-
-from fastapi import Depends, FastAPI, HTTPException
+from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
-from fastapi.security import OAuth2PasswordRequestForm
-from sqlalchemy import select
-from sqlalchemy.exc import IntegrityError
-from sqlalchemy.orm import Session
 
-from fast_zero.database import get_session
-from fast_zero.models import User
-from fast_zero.schemas import Message, Token, UserList, UserPublic, UserSchema
-from fast_zero.security import (
-    create_access_token,
-    get_current_user,
-    get_password_hash,
-    verify_password,
-)
+from fast_zero.routers import auth, users
+from fast_zero.schemas import Message
 
 app = FastAPI()
+
+app.include_router(auth.router)
+app.include_router(users.router)
 
 
 @app.get('/', response_model=Message)
@@ -28,140 +18,3 @@ def read_root():
 @app.get('/a2', response_class=HTMLResponse)
 def a2():
     return '<h1>ola mundo</h1>'
-
-
-@app.get(
-    '/users/{user_id}', status_code=HTTPStatus.OK, response_model=UserPublic
-)
-def read_user(
-    user_id: int,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    if current_user.id != user_id:
-        raise HTTPException(
-            detail='Not enough permission', status_code=HTTPStatus.FORBIDDEN
-        )
-
-    return current_user
-
-
-@app.get(
-    '/users/',
-    status_code=HTTPStatus.OK,
-    response_model=UserList,
-)
-def read_users(
-    offset: int = 0,
-    limit: int = 10,
-    session: Session = Depends(get_session),
-    current_user=Depends(get_current_user),
-):
-    users = session.scalars(select(User).offset(offset).limit(limit))
-    return {'users': users}
-
-
-@app.post('/users/', status_code=HTTPStatus.CREATED, response_model=UserPublic)
-def create_user(user: UserSchema, session: Session = Depends(get_session)):
-    db_user = session.scalar(
-        select(User).where(
-            (User.username == user.username) | (User.email == user.email)
-        )
-    )
-
-    if db_user:
-        if db_user.username == user.username:
-            raise HTTPException(
-                detail='Username already exist',
-                status_code=HTTPStatus.CONFLICT,
-            )
-
-        if db_user.email == user.email:
-            raise HTTPException(
-                detail='Email already exist', status_code=HTTPStatus.CONFLICT
-            )
-
-    db_user = User(
-        username=user.username,
-        email=user.email,
-        password=get_password_hash(user.password),
-    )
-
-    session.add(db_user)
-    session.commit()
-    session.refresh(db_user)
-
-    return db_user
-
-
-@app.put(
-    '/users/{user_id}', status_code=HTTPStatus.OK, response_model=UserPublic
-)
-def update_user(
-    user: UserSchema,
-    user_id: int,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    if current_user.id != user_id:
-        raise HTTPException(
-            detail='Not enough permission', status_code=HTTPStatus.FORBIDDEN
-        )
-
-    try:
-        current_user.username = user.username
-        current_user.email = user.email
-        current_user.password = get_password_hash(user.password)
-
-        session.add(current_user)
-        session.commit()
-        session.refresh(current_user)
-
-        return current_user
-    except IntegrityError:
-        raise HTTPException(
-            detail='Username or Email already exist',
-            status_code=HTTPStatus.CONFLICT,
-        )
-
-
-@app.delete(
-    '/users/{user_id}', status_code=HTTPStatus.OK, response_model=Message
-)
-def delete_user(
-    user_id: int,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    if current_user.id != user_id:
-        raise HTTPException(
-            detail='Not enough permission', status_code=HTTPStatus.FORBIDDEN
-        )
-
-    session.delete(current_user)
-    session.commit()
-
-    return {'message': 'User deleted'}
-
-
-@app.post('/token', response_model=Token)
-def login_for_access_token(
-    form_data: OAuth2PasswordRequestForm = Depends(),
-    session: Session = Depends(get_session),
-):
-    user = session.scalar(select(User).where(form_data.username == User.email))
-
-    if not user:
-        raise HTTPException(
-            datail='Incorrect email or password',
-            status_code=HTTPStatus.UNAUTHORIZED,
-        )
-
-    if not verify_password(form_data.password, user.password):
-        raise HTTPException(
-            datail='Incorrect email or password',
-            status_code=HTTPStatus.UNAUTHORIZED,
-        )
-
-    access_token = create_access_token({'sub': user.email})
-    return {'access_token': access_token, 'token_type': 'Bearer'}
