+++ pyproject.toml
@@ -33,7 +33,7 @@ extend-exclude = ['migrations']
 
 [tool.ruff.lint]
 preview = true
-select = ['I','F','E','W','PL','PT']
+select = ['I','F','E','W','PL','PT', 'FAST']
 
 [tool.ruff.format]
 preview = true
