+++ tests/test_app.py
@@ -1,7 +1,5 @@
 from http import HTTPStatus
 
-from fast_zero.schemas import UserPublic
-
 
 def test_root_return_hii(client):
     response = client.get('/')
@@ -15,170 +13,3 @@ def test_a2(client):
 
     assert response.text == '<h1>ola mundo</h1>'
     assert response.status_code == HTTPStatus.OK
-
-
-def test_create_user(client):
-    response = client.post(
-        '/users/',
-        json={
-            'username': 'alice',
-            'email': 'alice@example.com',
-            'password': 'alice@123',
-        },
-    )
-
-    assert response.status_code == HTTPStatus.CREATED
-    assert response.json() == {
-        'id': 1,
-        'username': 'alice',
-        'email': 'alice@example.com',
-    }
-
-
-def test_create_user_with_same_username(client, user):
-    response = client.post(
-        '/users/',
-        json={
-            'username': 'alice',
-            'email': 'alice@example.com',
-            'password': 'alice@123',
-        },
-    )
-
-    assert response.status_code == HTTPStatus.CONFLICT
-    assert response.json() == {'detail': 'Username already exist'}
-
-
-def test_create_user_with_same_email(client, user):
-    response = client.post(
-        '/users/',
-        json={
-            'username': 'bob',
-            'email': 'alice@example.com',
-            'password': 'alice@123',
-        },
-    )
-
-    assert response.status_code == HTTPStatus.CONFLICT
-    assert response.json() == {'detail': 'Email already exist'}
-
-
-def test_read_users(client, user, token):
-    user_schema = UserPublic.model_validate(user).model_dump()
-    response = client.get(
-        '/users/', headers={'Authorization': f'Bearer {token}'}
-    )
-
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {'users': [user_schema]}
-
-
-def test_update_user(client, user, token):
-    response = client.put(
-        '/users/1',
-        headers={'Authorization': f'Bearer {token}'},
-        json={
-            'username': 'bob',
-            'email': 'bob@example.com',
-            'password': 'bob@123',
-        },
-    )
-
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {
-        'id': 1,
-        'username': 'bob',
-        'email': 'bob@example.com',
-    }
-
-
-def test_update_another_user(client, user, token):
-    response = client.put(
-        f'/users/{user.id + 1}',
-        headers={'Authorization': f'Bearer {token}'},
-        json={
-            'username': 'angelico',
-            'email': 'angelico@example.com',
-            'password': 'angelico@123',
-        },
-    )
-
-    assert response.status_code == HTTPStatus.FORBIDDEN
-    assert response.json() == {'detail': 'Not enough permission'}
-
-
-def test_read_user(client, user, token):
-    response = client.get(
-        f'/users/{user.id}', headers={'Authorization': f'Bearer {token}'}
-    )
-
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {
-        'id': 1,
-        'username': 'alice',
-        'email': 'alice@example.com',
-    }
-
-
-def test_read_another_user(client, user, token):
-    response = client.get(
-        f'/users/{user.id + 1}',
-        headers={'Authorization': f'Bearer {token}'},
-    )
-
-    assert response.status_code == HTTPStatus.FORBIDDEN
-    assert response.json() == {'detail': 'Not enough permission'}
-
-
-def test_delete_user(client, user, token):
-    response = client.delete(
-        f'/users/{user.id}', headers={'Authorization': f'Bearer {token}'}
-    )
-
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {'message': 'User deleted'}
-
-
-def test_delete_another_user(client, user, token):
-    response = client.delete(
-        f'/users/{user.id + 1}', headers={'Authorization': f'Bearer {token}'}
-    )
-
-    assert response.status_code == HTTPStatus.FORBIDDEN
-    assert response.json() == {'detail': 'Not enough permission'}
-
-
-def test_update_integrity_error(client, user, token):
-    client.post(
-        '/users',
-        json={
-            'username': 'Jhon Doe',
-            'email': 'jhon@email.com',
-            'password': 'jhon@123',
-        },
-    )
-
-    response = client.put(
-        f'users/{user.id}',
-        headers={'Authorization': f'Bearer {token}'},
-        json={
-            'username': 'Jhon Doe',
-            'email': 'jhon@email.com',
-            'password': 'jhon@123',
-        },
-    )
-
-    assert response.status_code == HTTPStatus.CONFLICT
-    assert response.json() == {'detail': 'Username or Email already exist'}
-
-
-def test_get_token(client, user):
-    response = client.post(
-        '/token',
-        data={'username': user.email, 'password': user.clean_password},
-    )
-
-    token = response.json()
-    assert response.status_code == HTTPStatus.OK
-    assert token['token_type'] == 'Bearer'
-    assert 'access_token' in token
