+++ tests/test_security.py
@@ -2,14 +2,14 @@ from http import HTTPStatus
 
 from jwt import decode
 
-from fast_zero.security import ALGORITHM, SECRET_KEY, create_access_token
+from fast_zero.security import create_access_token
 
 
-def test_jwt():
+def test_jwt(settings):
     data = {'test': 'test'}
 
     token = create_access_token(data)
-    decoded = decode(token, SECRET_KEY, ALGORITHM)
+    decoded = decode(token, settings.SECRET_KEY, settings.ALGORITHM)
 
     assert decoded['test'] == data['test']
     assert 'exp' in decoded
